src/azure-cli/requirements.py3.Linux.txt
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
argcomplete==1.11.1
asn1crypto==0.24.0
azure-appconfiguration==1.1.1
azure-batch==11.0.0
azure-cli-core==2.33.0
azure-cli-telemetry==1.0.6
azure-cli==2.33.0
azure-common==1.1.22
azure-core==1.21.1
azure-cosmos==3.2.0
azure-datalake-store==0.0.49
azure-graphrbac==0.60.0
azure-identity==1.6.1
azure-keyvault-administration==4.0.0b3
azure-keyvault-keys==4.5.0b4
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==16.0.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==0.3.0
azure-mgmt-cdn==11.0.0
azure-mgmt-cognitiveservices==13.0.0
azure-mgmt-compute==23.1.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==9.1.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-containerservice==16.4.0
azure-mgmt-core==1.3.0
azure-mgmt-cosmosdb==7.0.0b2
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==9.0.0
azure-mgmt-eventhub==9.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.0.0
azure-mgmt-iotcentral==9.0.0
azure-mgmt-iothub==2.1.0
azure-mgmt-iothubprovisioningservices==1.0.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==12.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==7.0.0
azure-mgmt-monitor==3.0.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==6.0.1
azure-mgmt-network==19.3.0
azure-mgmt-policyinsights==1.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.0.0
azure-mgmt-recoveryservices==2.0.0
azure-mgmt-recoveryservicesbackup==4.0.0
azure-mgmt-redhatopenshift==1.0.0
azure-mgmt-redis==13.1.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==20.0.0
azure-mgmt-search==8.0.0
azure-mgmt-security==2.0.0b1
azure-mgmt-servicebus==6.0.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-servicelinker==1.0.0b1
azure-mgmt-signalr==1.0.0b2
azure-mgmt-sql==3.0.1
azure-mgmt-sqlvirtualmachine==1.0.0b1
azure-mgmt-storage==19.0.0
azure-mgmt-synapse==2.1.0b2
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==4.0.0
azure-multiapi-storage==0.7.0
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.10.0
azure-synapse-managedprivateendpoints==0.3.0
azure-synapse-spark==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.2
distro==1.6.0
fabric==2.4.0
humanfriendly==10.0
idna==2.8
invoke==1.2.0
isodate==0.6.0
javaproperties==0.5.1
Jinja2==2.11.3
jmespath==0.9.5
jsondiff==1.3.0
knack==0.9.0
MarkupSafe==1.1.1
msal-extensions==0.3.1
msal==1.16.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.0.1
packaging==21.3
paramiko==2.6.0
pbr==5.3.1
portalocker==2.3.2
psutil==5.9.0
pycparser==2.19
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
requests-oauthlib==1.2.0
requests[socks]==2.26.0
scp==0.13.2
semver==2.13.0
six==1.14.0
sshtunnel==0.1.5
tabulate==0.8.9
urllib3==1.26.7
vsts==0.1.25
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.2
xmltodict==0.12.0
